requirements.txt
asgiref==3.8.1
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
filelock==3.13.3
gunicorn==21.2.0
identify==2.5.35
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
uritemplate==4.1.1
virtualenv==20.25.1
djoser==2.2.2
